Keep fire from getting out of control
The prescribed method for putting out the fire is fireplace-off,
which is assigned to q in fireplace-mode. But personally I'm in the
habit of just killing buffers when I'm done with them. This is a big
problem for fireplace, because killing the buffer fails to extinguish
all the embers, and those embers start new fires. That is, killing the
buffer fails to cancel the draw timer, and the draw timer will then
continue to create *fireplace* buffers and draw flames. What's worse,
these new buffers won't be in fireplace-mode, so it's not even
possible to put out the flames with q! This problem was one of two
mentioned in https://github.com/johanvts/emacs-fireplace/issues/3.

This commit adds a fireplace-mode check in fireplace-draw. Say I start
fireplace and then kill the buffer directly, without running
fireplace-off. The draw timer will continue to run, and it will create
a new *fireplace* buffer in fundamental-mode. Then the fireplace-mode
check will run and determine that the buffer is no good, and the
flames will put out themselves. This is a somewhat hacky solution, but
since this is a piece of novelty software, I can live with it.