Fix: Opening up a new fireplace no longer enables transient-mark-mode
There was a problem that was driving me up the wall after I tinkered with
my fireplace config, and I didn't realize it for ages,
but when you call 'fireplace' it also calls 'fireplace--disable-minor-modes',
which, I think it rather innocently tries to disable some functions to help
improve the visual experience, including 'transient-mark-mode'.
Actually it *enables* it.

The mode system in emacs is quirky and when using a variable, nil is off and
non-nil is on.
But for mode functions, generally negative arguments is off, and non-nil
or omitted arguments is on.
So, paradoxically, the omitted 'nil', enables it, which also happens to be a
global mode.

The replacement tries to fix it by using setq-local instead to prevent global
settings from being overridden.
I have no idea if it will cause issues with not affecting the modes on the
fireplace start, as that often happens with settings variables and modes with
variables, especially locally.
But its much better than modifying the global environment and hoping it does not
affect the user.